update filters:
consolidate filter

consolidate filter

save work

update options

delete company typeahead component since we genericized it with asynctypeaheadwrapper

remove wrapper and use typeahead directly

remove handleFetchSearch

fixing unit test

fixing nested

fixing test

fix test, lint

fixing tests

update tests

update dist

adding MSA and Cong Dist filters to frontend

remove zipcode filter

update dist
